Spreads event update
Change the spread event update by accounting for changing favorites.
Currently, spread points are subtracted from the favorite's score, and the favorite is fixed as soon as it is deduced from the initial event odds.
However, favorites can change during an event, and consequently, spread points should be subtracted from the current favorite (which might be the old underdog).
This update introduces a mechanism to account spread points without looking at the initial favorite.

Note that the commit highlights several issues with the current implementation, that need updating/optimizing, including:
- Logic to handle odds and results is scattered over the code base.
 - Variables related to results are set in sections that manage event data and vice versa
 - Odds are set in various sections.
- Variables of the wrong type are used. For example: most event data is stored in uint32_t, while only some of them are actually hold 32bit unsigned numbers.
- Event data is not taken from databases but scanned from the block files
- The version numbers from the event and bet types are not stored in memory, not stored in the database and not used beyond a check if they are 8 bits.
- There is no functional test or unit test that can be extended to verify that spreads in the new style will be paid out correctly (functional) or that the opcodes are parsed correctly (unit).