Rework staking status
This PR cleans up the miner code and streamlines the logic used by the client (both in the RPC and in the GUI) to determine whether the wallet is staking or not.
Currently it relies on a collection of variables (mapHashedBlocks, nLastCoinStakeInterval, nSearchInterval, etc...) updated in kernel.
These are here replaced with a new class CStakerStatus, updated in CreateCoinStake (thus directly in wallet), and an instance of CStakerStatus is introduced as member of CWallet.
This class keeps two variables, timeLastStakeAttempt and tipLastStakeAttempt, containing respectively the time of last stake attempt and a pointer to the index of the block upon which last stake attempt was made.

The Staking Status is Active whenever timeLastStakeAttempt is less than
30 seconds in the past.

This PR also expands the output of getstakingstatus rpc call adding:

staking_enabled to tell whether or not staking has been disabled via conf file / startup flag
hashLastStakeAttempt and timeLastStakeAttempt values
heightLastStakeAttempt (the height of the block with hash hashLastStakeAttempt)
tiptime (the time of the chaintip block) to compare with timeLastStakeAttempt. This field replaces valid_time which is unnecessary now).
and fixes enoughcoins with proper staking balance.